requirements.txt
click
httpx
loguru
nuitka
pydantic
pyyaml
zstandard